+++ docs/source/changelog.rst
@@ -10,6 +10,10 @@ ChangeLog
 1.23.0 (TBR)
 ************
 
+* The profile used when issuing certificates via ACMEv2 is now configurable by certificate authority. The
+  default is the profile named in :ref:`settings-ca-default-profile`, it was previously hardcoded to use the
+  "server" profile.
+
 Backwards incompatible changes
 ==============================